refactor: split cache.py into operations modules
Split monolithic 60K cache.py into focused operation modules:
- operations/lookup.py - search and matching
- operations/storage_ops.py - store and eviction
- operations/invalidation.py - deletion and cleanup
- operations/maintenance.py - stats and I/O

cache.py now acts as facade, all public APIs unchanged.
No breaking changes, all tests pass.
